requirements.txt
flask
python-dotenv
openai
werkzeug
torch
git+https://github.com/openai/whisper.git
numpy
ffmpeg-python 